requirement.txt
flask
numpy
tensorflow
pillow
waitress